requirements.txt
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
tenacity>=8.2.0
langdetect>=1.0.9
rapidfuzz>=3.5.0
ruff>=0.9.0
mypy>=1.8.0
pytest>=8.0.0
types-requests>=2.31.0
